issue #58 - Add function to delete author (#82)

+++ database/core/functions/delete_book.sql
@@ -16,6 +16,19 @@ begin
     );
   end if;
 
+  if not exists(select 1
+                from core.books b
+                where b.id = _book_id)
+  then
+    return jsonb_build_object(
+      'status', 1,
+      'details', jsonb_build_object(
+        'message', 'Book not found.',
+        'code', 'NOT_FOUND'
+      )
+    );
+  end if;
+
   delete from main.books b
   where b.id = _book_id;
